+++ Enamel/Systems/SelectionPreviewSystem.cs
@@ -138,7 +138,7 @@ public class SelectionPreviewSystem : SpellSystem
     {
         var preview = _world.CreateEntity();
         Set(preview, new GridCoordComponent(x, y));
-        Set(preview, new TextureIndexComponent((int) Sprite.SelectPreview));
+        Set(preview, new TextureIndexComponent(Sprite.SelectPreview));
 
         return preview;
     }
